meme/pyproject.toml
[tool.poetry]
name = "meme"
version = "0.1.0"
description = ""
authors = ["WATUNeed <maxconal228@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
faststream = {version = "^0.5.14", extras = ["rabbit"]}
pydantic-settings = "^2.3.4"
sqlalchemy = "^2.0.31"
alembic = {extras = ["asyncio"], version = "^1.13.1"}
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.7"
pytest-dotenv = "^0.5.2"
pytest-mock = "^3.14.0"
pytest-env = "^1.1.3"
faker = "^25.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
